==> analyses/preclinical/02.process-cell-line-molecular-features.ipynb <==
# Format molecular features of cell lines for use by Molecular Oncology Almanac
We format somatic variants, copy nubmer alterations, and fusions for use by the Molecular Oncology Almanac to test for preclinical efficacy and for matchmaking. For each data type, we will create a table that contains at most five columns:
- `model_id`: The cell line sample name
- `feature_type`: The feature type (e.g. Somatic Variant, Copy Number, Rearrangement)
- `feature`: The gene of interest
- `alteration_type`: Variant classification or copy number direction
- `alteration`: Protein change or fusion partner

Additionally, we will import GDSC sensitivity data to create an "about" table that contains the following for each cell line studied:
- booleans for each data type 
- additional names of the cell line 
- cell line ontology

We use cell lines that meet the following criteria:
- map to a CCLE name
- have all four data types available
- is not a blood cancer 
- is not subject to genetic drift
- is not contaminated by fibroblast
- for evaluation, they must be sensitive to at least one therapy and there must be at least one other cell line that shares therapeutic sensitivity

In all, we are left with 377 cell lines for evaluation and 452 for production.

import pandas as pd

gdsc_model_info = pd.read_csv('source/gdsc/model_list_20200204.csv')
ccle_model_info = pd.read_csv('source/ccle-2019/data_clinical_sample.txt', sep='\t', comment='#')

fibroblast_maps = (pd.
                   read_excel('source/ccle-2019/41586_2019_1186_MOESM4_ESM.xlsx', 
                              sheet_name='Cell line name changes')
                   .iloc[:45, :]
                   .set_index('old_CCLE_ID')
                   .loc[:, 'new_CCLE_ID']
                   .to_dict()
                  )

summary = pd.read_csv('formatted/cell-line-names.formatted.txt', sep='\t')
summary['fibroblast'] = (summary['ccle_name'].isin(fibroblast_maps.keys()) | 
                         summary['alt_ccle'].isin(fibroblast_maps.keys()) | 
                         summary['ccle_name'].isin(fibroblast_maps.values()) |
                         summary['alt_ccle'].isin(fibroblast_maps.values())
                        ).astype(int)

drift = pd.read_excel('source/ccle-2019/41586_2019_1186_MOESM6_ESM.xlsx', sheet_name='CCLE_GDSC_comparison')
drift_samples = drift[drift['r_somatic_CCLE_HC_vs_GDSC_WES'].fillna(0).lt(0.75)]['CCLE_ID']
summary['drift'] = (summary['ccle_name'].isin(drift_samples) | summary['alt_ccle'].isin(drift_samples)).astype(int)

ccle_alt_map = (summary
 .loc[summary['alt_ccle'].dropna().index, ['ccle_name', 'alt_ccle']]
 .set_index('alt_ccle')
 .loc[:, 'ccle_name']
 .to_dict()
)

summary['fibroblast'].value_counts()

summary['drift'].value_counts()

## Somatic Variants

handle = 'source/ccle-2019/data_mutations_extended.txt'
df = pd.read_csv(handle, sep='\t', comment='#', low_memory=False)
df['Tumor_Sample_Barcode'].replace(fibroblast_maps, inplace=True)
df['Tumor_Sample_Barcode'].replace(ccle_alt_map, inplace=True)
somatic_samples = df['Tumor_Sample_Barcode'].drop_duplicates().sort_values().tolist()

df = df[~df['Chromosome'].eq('M')]

classification_map = {
    'Missense_Mutation': 'Missense',
    'Frame_Shift_Del': 'Frameshift',
    'Frame_Shift_Ins': 'Frameshift',
    'Nonsense_Mutation': 'Nonsense',
    'Splice_Site': 'Splice Site',
    'In_Frame_Ins': 'Insertion',
    'In_Frame_Del': 'Deletion',
    'Nonstop_Mutation': 'Nonstop'
}

idx_coding = df['Variant_Classification'].isin(list(classification_map.keys()))
silent = df[df['Variant_Classification'].eq('Silent')]

df = df[idx_coding]
df['Variant_Classification'] = df['Variant_Classification'].replace(classification_map)

somatic = df.loc[:, ['Tumor_Sample_Barcode', 'Hugo_Symbol', 'Variant_Classification', 'Protein_Change']]
somatic = somatic.rename(columns = {'Hugo_Symbol': 'feature', 
                                   'Variant_Classification': 'alteration_type',
                                   'Protein_Change': 'alteration',
                                   'Tumor_Sample_Barcode': 'model_id'})
silent.rename(columns={'Tumor_Sample_Barcode': 'model_id'}, inplace=True)

somatic = (somatic
           .loc[:, ['feature', 'alteration_type', 'alteration', 'model_id']]
           .sort_values(['model_id', 'feature', 'alteration_type', 'alteration'])
           .drop_duplicates()
           .reset_index(drop=True)
          )

somatic.to_csv('formatted/ccle.variants.txt', sep='\t', index=False)
somatic.head()

summary['somatic_variants'] = (summary['ccle_name'].isin(somatic_samples) | 
                               summary['alt_ccle'].isin(somatic_samples)
                              ).astype(int)
summary['somatic_variants'].value_counts()

somatic['model_id'].value_counts().describe()

for gene in ['EGFR', 'CDKN2A', 'BRAF']:
    print(somatic[somatic['feature'].eq(gene)].shape[0] / len(somatic_samples))

## Copy Number

handle = 'source/ccle-2019/data_CNA.txt'
df = pd.read_csv(handle, sep='\t', low_memory=False)
df = df.set_index('Hugo_Symbol')
df.rename(columns=fibroblast_maps, inplace=True)
df.rename(columns=ccle_alt_map, inplace=True)

copy_number_samples = df.columns.tolist()

df = df.stack().to_frame().reset_index()
df[0] = df[0].astype(float)
df = df[df[0].abs().ge(2)]
df['mapped'] = df[0].replace({2.0: 'Amplification', 
                              -2.0: 'Deletion', 
                              1.0: 'Amplification',
                              -1.0: 'Deletion'})
df = df.rename(columns = 
          {'Hugo_Symbol': 'feature',
           'level_1': 'model_id',
           'mapped': 'alteration_type'
          })
df['alteration'] = ''

copy_number = (df
               .loc[:, ['feature', 'alteration_type', 'alteration', 'model_id']]
               .sort_values(['model_id', 'feature', 'alteration_type', 'alteration'])
               .drop_duplicates()
               .reset_index(drop=True)
     )

copy_number.to_csv('formatted/ccle.copy-numbers.txt', sep='\t', index=False)

summary['copy_number_alterations'] = (summary['ccle_name'].isin(copy_number_samples) | 
                                      summary['alt_ccle'].isin(copy_number_samples)
                                     ).astype(int)
summary['copy_number_alterations'].value_counts()

for gene in ['EGFR', 'CDKN2A', 'BRAF']:
    print(copy_number[copy_number['feature'].eq(gene)].shape[0] / len(copy_number_samples))

copy_number['model_id'].value_counts().describe()

copy_number['alteration_type'].value_counts()

## Fusion

handle = 'source/gdsc/fusions_20191101.csv'
df = pd.read_csv(handle)

fusion_samples = df['model_id'].drop_duplicates().sort_values().tolist()

df = (df
      .rename(columns={'gene_symbol_3prime': 'feature', 'gene_symbol_5prime': 'partner'})
      .loc[:, ['feature', 'partner', 'model_id']]
      .drop_duplicates()
)

fusions = (df
          .sort_values(['model_id', 'feature', 'partner'])
          .drop_duplicates(['model_id', 'feature', 'partner'])
          .reset_index(drop=True)
         )

fusions.to_csv('formatted/sanger.fusions.txt', sep='\t', index=False)
fusions.head()

summary['fusions'] = (summary['sanger'].isin(fusion_samples)).astype(int)
summary['fusions'].value_counts()

len(fusion_samples)

## Therapeutic sensitivity

gdsc1 = pd.read_excel('source/gdsc/GDSC1_fitted_dose_response_25Feb20.xlsx')
gdsc2 = pd.read_excel('source/gdsc/GDSC2_fitted_dose_response_25Feb20.xlsx')
gdsc = pd.concat([
    gdsc1[~gdsc1['DRUG_NAME'].isin(gdsc2['DRUG_NAME'])],
    gdsc2
], ignore_index=True)

gdsc_samples = gdsc['SANGER_MODEL_ID'].drop_duplicates().sort_values()

gdsc = (gdsc
        .loc[:, ['DATASET', 'SANGER_MODEL_ID', 'DRUG_NAME', 'LN_IC50', 'AUC', 'Z_SCORE']]
        .rename(columns={'SANGER_MODEL_ID': 'model_id', 'DRUG_NAME': 'therapy_name', 
                         'LN_IC50': 'ln_ic50', 'Z_SCORE': 'z_score'})
       )

gdsc = gdsc.sort_values('z_score', ascending=True).drop_duplicates(['model_id', 'therapy_name'], keep='first')

gdsc.to_csv('formatted/sanger.gdsc.txt', sep='\t', index=False)
gdsc.head()

summary['gdsc'] = (summary['sanger'].isin(gdsc_samples)).astype(int)
summary['gdsc'].value_counts()

len(gdsc_samples)

## Summary

output = summary.copy(deep=True).reset_index(drop=True)
output['n_dtypes'] = output.loc[:, ['somatic_variants', 'copy_number_alterations', 'fusions', 'gdsc']].sum(axis=1)
output = output.merge(ccle_model_info
                      .rename(columns={'SAMPLE_ID': 'ccle_name'})
                      .loc[:, ['ccle_name', 'PRIMARY_SITE', 'TUMOR_TYPE', 'TYPE_REFINED']], 
                      on='ccle_name', how='left')

output = output.merge(gdsc_model_info
                      .rename(columns={'model_id': 'sanger', 
                                       'cancer_type': 'sanger_cancer_type', 
                                       'cancer_type_detail': 'sanger_cancer_type_detail'})
                      .loc[:, ['sanger', 'sanger_cancer_type', 'sanger_cancer_type_detail']],
                      on='sanger', how='left')

series = pd.Series(index=output['sanger'].dropna().tolist(), dtype=int)
for label, group in gdsc[gdsc['z_score'].le(-2)].groupby('model_id'):
    series.loc[label] = group.shape[0]
series.fillna(0, inplace=True)

output = (output
           .merge(
               series
               .reset_index()
               .rename(columns={'index': 'sanger', 0: 'n_therapies_sensitive'}),
               on='sanger', how='left'
    )
)    
output['n_therapies_sensitive'].fillna(pd.NA, inplace=True)

output = (output
           .merge(
               silent['model_id']
               .value_counts()
               .reset_index()
               .rename(columns={'index': 'ccle_name', 'model_id': 'silent_var_counts'}),
               on='ccle_name', how='left'
           )
)
output['silent_var_counts'].fillna(pd.NA, inplace=True)

idx_blood = (output['PRIMARY_SITE'].str.contains('Haematopoietic') | 
             output['ccle_name'].str.contains('HAEMATOPOIETIC') |
             output['sanger_cancer_type'].str.contains('Leukemia') | 
             output['sanger_cancer_type'].str.contains('Other Blood Carcinomas') | 
             output['sanger_cancer_type'].str.contains('Myeloma')
            ).fillna(False)
idx_drift = output['drift'].fillna(0).eq(1)
idx_fibroblast = output['fibroblast'].fillna(0).eq(1)
idx_reject = (idx_blood | idx_fibroblast | idx_drift)
idx_sensitive_one_therapy = output['n_therapies_sensitive'].fillna(0).ge(1)
idx_4_dtypes = output['n_dtypes'].eq(4)

output['use_evaluate'] = idx_sensitive_one_therapy & idx_4_dtypes & ~idx_reject
output['use'] = idx_4_dtypes & ~idx_reject

for column in ['use_evaluate', 'use']:
    print(column)
    print(output[column].value_counts())
    print('')
    
#output.to_csv('formatted/cell-lines.summary.txt', sep='\t', index=False)

output[output['use'].eq(1)]['ccle_name'].isnull().value_counts()

summary.head()

## Create shared sensitivity dictionary
We load metadata on cell lines and their sensitivity data and create a table of pairwise comparisons, counting the number of therapies that both are sensitive to. The output is saved to `formatted/sanger.gdsc.pairwise-sensitive.txt`.

We also need to update our evaluate to require that there is at least one other sample that shares therapeutic sensitivity, not counting itself.

We find that there are two samples used for evaluation that do not share a therapeutic sensitivity with other samples, so we remove them from the evaluation pool.

from sklearn import neighbors, metrics

### Load GDSC and compare drug similarity
1. Create boolean table
2. Create distance table
3. Create stacked table

zscore_cutoff = -2
therapies = gdsc['therapy_name'].drop_duplicates().sort_values().tolist()
therapy_dict = {}
for label, group in gdsc.groupby('model_id'):
    sensitive_therapies = group[group['z_score'].le(zscore_cutoff)]['therapy_name'].tolist()
    therapy_dict[label] = sensitive_therapies
len(list(therapy_dict.keys()))

### Intersection size
The naive thing to do is how many therapies sample A is sensitive to and how many therapies sample B is sensitive to. Then, calculate the fraction. One issue with this is that not every sample is treated with the same therapies and so there may be differential sensitivity simply because another sample was not tested with a drug. 

Instead, for every therapy, we need to characterize:
- the sample name (`model_id`)
- the number of therapies tested (`n_tested`)
- the number of therapies sensitive to (`n_sensitve`)
- the therapies tested (`tested`)
- the therapies sensitive to (`sensitive`)

Then, for every case comparison pair, 
- the name of case (`case`)
- the name of comparison (`comparison`)
- the name of therapies tested, case (`tested_case`)
- the name of therapies tested, comparison (`tested_comparison`)
- the name of therapies sensitive to, case (`sensitive_case`)
- the name of therapies sensitive to, comparison (`sensitive_comparison`)
- the union of sensitive therapies (`union_sensitive`)
- the number of therapies with shared sensitivity (`n_union_sensitive`)

case_dict = {}
for label, group in gdsc.groupby('model_id'):
    tested_therapies = group['therapy_name'].sort_values().tolist()
    sensitive_therapies = group[group['z_score'].le(zscore_cutoff)]['therapy_name'].sort_values().tolist()
    case_dict[label] = {} 
    case_dict[label]['n_tested'] = len(tested_therapies)
    case_dict[label]['n_sensitive'] = len(sensitive_therapies)
    case_dict[label]['tested'] = tested_therapies
    case_dict[label]['sensitive'] = sensitive_therapies
    
def list_intersection(list1, list2):
    return [value for value in list1 if value in list2]
import time

samples = gdsc['model_id'].drop_duplicates().sort_values().tolist()

list_of_series = []
for case in samples:
    for comparison in samples:
        series = pd.Series(dtype=str)
        series.loc['case'] = case
        series.loc['comparison'] = comparison
        
        case_sensitive = case_dict[case]['sensitive']
        comparison_sensitive = case_dict[comparison]['sensitive']
        union_sensitive = list_intersection(case_sensitive, comparison_sensitive)
        union_sensitive_str = ', '.join(union_sensitive)
        
        series.loc['n_sensitive_case'] = len(case_sensitive)
        series.loc['n_sensitive_comparion'] = len(comparison_sensitive)
        series.loc['n_sensitive_union'] = len(union_sensitive)
        series.loc['union_sensitive'] = union_sensitive_str
        list_of_series.append(series)

df = pd.concat(list_of_series, axis=1).T
df.to_csv('formatted/sanger.gdsc.pairwise-sensitive.txt', sep='\t', index=False)

samples_evaluate = output[output['use_evaluate'].eq(True)]['sanger'].tolist()
counts_of_sensitive_neighbors = pd.Series(index=samples_evaluate)
df_for_evaluation = df[df['case'].isin(samples_evaluate) & df['comparison'].isin(samples_evaluate)]

for label, group in df_for_evaluation.groupby('case'):
    tmp = group[~group['comparison'].eq(label) & group['n_sensitive_union'].gt(0)]
    n = tmp.shape[0]
    counts_of_sensitive_neighbors.loc[label] = n

counts_of_sensitive_neighbors[counts_of_sensitive_neighbors.eq(0)]

for sanger_id in counts_of_sensitive_neighbors[counts_of_sensitive_neighbors.eq(0)].index.tolist():
    index = output[output['sanger'].eq(sanger_id)].index
    output.loc[index, 'use_evaluate'] = False

for column in ['use_evaluate', 'use']:
    print(column)
    print(output[column].value_counts())
    print('')
    
output.to_csv('formatted/cell-lines.summary.txt', sep='\t', index=False)
