==> analyses/preclinical/08.create-model-performance-tables.ipynb <==
import numpy as np
import pandas as pd
import pickle

def read_pickle(handle):
    return pickle.load(open(handle, "rb"))

models = read_pickle('tables/models.evaluated.pkl')
models_random = read_pickle('tables/models.random.evaluated.pkl')

def summarize_models(db, random):
    columns = ['model_name', 
               'mean_average_precision', 
               'average_precision_at_k=1', 
               'average_precision_at_k=2', 
               'average_precision_at_k=3', 
               'average_precision_at_k=4',
               'average_precision_at_k=5',
               'description']

    list_of_series = []
    for model in db.keys():
        series = pd.Series(pd.NA, name=model, index=columns)
        series.loc['model_name'] = model
        series.loc['mean_average_precision'] = db[model]['mean_average_precision']
        for k in range(1, 6):
            series.loc['average_precision_at_k={}'.format(k)] = db[model]['ap@k'][k]
        series.loc['description'] = db[model]['description']
        list_of_series.append(series.to_frame().T)
    
    for model in random.keys():
        series = pd.Series(pd.NA, name=model, index=columns)
        series.loc['model_name'] = model
        series.loc['mean_average_precision'] = random[model]['mean_average_precision']
        for k in range(1, 6):
            series.loc['average_precision_at_k={}'.format(k)] = random[model]['ap@k'][k]
        series.loc['description'] = random[model]['description']
        list_of_series.append(series.to_frame().T)
    
    
    return pd.concat(list_of_series, ignore_index=True)

normal = summarize_models(models, models_random)
normal.sort_values(['average_precision_at_k=1', 'mean_average_precision'], ascending=False, inplace=True)
normal.to_csv('tables/models.performance.txt', sep='\t', index=False)

(normal
 .loc[:, ['model_name', 'average_precision_at_k=1', 'mean_average_precision', 'description']]
 .head(5)
)

model_performance = pd.read_csv('tables/models.performance.txt', sep='\t')
pairwise_performance = pd.read_csv('tables/pairwise-model-comparison.txt', sep='\t')

folder = '../../tables/supplementary-table-07'
filename = 'supplementary-table-7'
!mkdir -p $folder

with pd.ExcelWriter(f'{folder}/{filename}.xlsx') as writer:  
    model_performance.to_excel(writer, sheet_name='model perfromance', index=False)
    pairwise_performance.to_excel(writer, sheet_name='pairwise model comparison', index=False)
