==> analyses/preclinical/11.nearest-neighbors-prospective-cohort.ipynb <==
# Nearest neighbors from prospective cohorts
We applied matchmaking to the prospective clinical trial (I-PREDICT). In this notebook, we look at the results; specifically....

import glob
import matplotlib.gridspec as gridspec
import matplotlib.pyplot as plt
import pandas as pd

import sys
sys.path.append('../../common/')
import settings
settings.set_arial_as_font()

from settings import Colors
tableau10 = Colors.tableau10

%matplotlib inline

handles = glob.glob('../prospective-trial/almanac_outputs/*/*.matchmaker.txt')

summary = pd.read_csv('formatted/cell-lines.summary.txt', sep='\t')
zscores = pd.read_csv('formatted/sanger.gdsc.txt', sep='\t')

df = pd.concat([pd.read_csv(handle, sep='\t', index_col=0) for handle in handles])
df = df[~df['comparison'].eq('case-profile')]
df.sort_values(['SNF: FDA & CGC'], ascending=True, inplace=True)
df.reset_index(inplace=True)
df['case'] = df['case'].astype(str)

broad_to_ccle_map = summary.set_index('broad')['ccle_name'].dropna().to_dict()
sanger_to_ccle_map = summary.set_index('sanger')['ccle_name'].dropna().to_dict()

zscores['ccle_name'] = zscores['model_id'].replace(sanger_to_ccle_map)

first_neighbors = pd.Series(index=df['case'].astype(str).sort_values().drop_duplicates())
for label, group in df.groupby('case'):
    nearest = group['comparison'].replace(broad_to_ccle_map).tolist()[0]
    first_neighbors.loc[label] = nearest
first_neighbors.name = 'nearest neighbor'
first_neighbors = first_neighbors.to_frame().reset_index()


for index in first_neighbors.index: 
    cell_line = first_neighbors.loc[index, 'nearest neighbor']
    therapies = zscores[zscores['ccle_name'].eq(cell_line) & zscores['z_score'].le(-2)]['therapy_name'].tolist()
    n = len(therapies)
    first_neighbors.loc[index, 'sensitive therapies'] = ', '.join(therapies)
    first_neighbors.loc[index, 'n'] = n

first_neighbors['n'].value_counts()

first_neighbors['n'].sum()

first_neighbors['n'].describe()

first_neighbors['case'] = first_neighbors['case'].astype(str)
(first_neighbors
 .sort_values(['case'])
 .to_csv('tables/ipredict/ipredict.nearest_neighbors.txt', sep='\t', index=False)
)

first_neighbors
